backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pymupdf>=1.24.0
faster-whisper>=1.0.3
langgraph>=0.2.0
openai>=1.30.0
pgvector>=0.3.0
psycopg2-binary>=2.9.9
sentence-transformers>=3.0.0
torch>=2.3.0
torchvision>=0.18.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
numpy>=1.26.0
Pillow>=10.0.0
ultralytics>=8.0.0
